@@ -1,32 +1,48 @@
 {
   "name": "vieew",
-  "main": "src/main.ts",
-  "version": "1.0.0",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
+  },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~15.2.0",
-    "@angular/common": "~15.2.0",
-    "@angular/compiler": "~15.2.0",
-    "@angular/core": "~15.2.0",
-    "@angular/forms": "~15.2.0",
-    "@angular/platform-browser": "~15.2.0",
-    "@angular/platform-browser-dynamic": "~15.2.0",
-    "@angular/router": "~15.2.0",
-    "@nativescript/angular": "^15.0.0",
-    "@nativescript/core": "~8.5.0",
-    "@nativescript/theme": "~3.0.2",
-    "nativescript-ui-listview": "^15.2.3",
-    "nativescript-ui-sidedrawer": "~15.2.0",
-    "rxjs": "~7.6.0",
-    "zone.js": "~0.13.0"
+    "@angular/animations": "^15.2.0",
+    "@angular/common": "^15.2.0",
+    "@angular/compiler": "^15.2.0",
+    "@angular/core": "^15.2.0",
+    "@angular/forms": "^15.2.0",
+    "@angular/platform-browser": "^15.2.0",
+    "@angular/platform-browser-dynamic": "^15.2.0",
+    "@angular/router": "^15.2.0",
+    "@aws-amplify/ui-angular": "^3.2.18",
+    "@ng-bootstrap/ng-bootstrap": "^14.2.0",
+    "@popperjs/core": "^2.11.6",
+    "aws-amplify": "^5.2.4",
+    "bootstrap": "^5.2.3",
+    "camelcase-keys": "^8.0.2",
+    "jwt-decode": "^3.1.2",
+    "rxjs": "~7.8.0",
+    "snakecase-keys": "^5.4.5",
+    "sweetalert2": "^11.7.10",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~15.2.0",
-    "@angular/compiler-cli": "~15.2.0",
-    "@nativescript/android": "8.5.0",
-    "@nativescript/types": "~8.5.0",
-    "@nativescript/webpack": "~5.0.14",
-    "@ngtools/webpack": "~15.2.0",
-    "typescript": "~4.8.4"
+    "@angular-devkit/build-angular": "^15.2.6",
+    "@angular/cli": "~15.2.6",
+    "@angular/compiler-cli": "^15.2.0",
+    "@angular/localize": "^15.2.0",
+    "@types/jasmine": "~4.3.0",
+    "jasmine-core": "~4.5.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
+    "typescript": "~4.9.4"
   }
 }
