@@ -1,25 +1,35 @@
-import { NgModule, NO_ERRORS_SCHEMA } from '@angular/core'
+import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+
 // Modules
-import { SharedModule } from './shared/shared.module'
-import { CoreModule } from './shared/core.module'
-import { AppRoutingModule } from './app-routing.module'
+import { AmplifyAuthenticatorModule } from '@aws-amplify/ui-angular';
+import { SharedModule } from './shared/shared.module';
+import { CoreModule } from './shared/core.module';
+
+import { AppRoutingModule } from './app-routing.module';
 // Components
-import { AppComponent } from './app.component'
-import { LoginComponent } from './shared/components/login/login.component'
-import { RegisterComponent } from './shared/components/register/register.component'
+import { AppComponent } from './base/app.component';
+import { WelcomeComponent } from './base/components/welcome/welcome.component';
+import { LoginComponent } from './base/components/login/login.component';
+import { RegisterComponent } from './base/components/register/register.component';
+import { SecuredLayoutComponent } from './base/components/secured-layout/secured-layout.component';
 
 @NgModule({
-  bootstrap: [AppComponent],
-  imports: [
-    AppRoutingModule,
-    CoreModule,
-    SharedModule
-  ],
   declarations: [
     AppComponent,
+    WelcomeComponent,
     LoginComponent,
-    RegisterComponent
+    RegisterComponent,
+    SecuredLayoutComponent
   ],
-  schemas: [NO_ERRORS_SCHEMA],
+  imports: [
+    BrowserModule,
+    AppRoutingModule,
+    CoreModule,
+    SharedModule,
+    AmplifyAuthenticatorModule,
+  ],
+  providers: [],
+  bootstrap: [AppComponent]
 })
-export class AppModule {}
+export class AppModule { }
