@@ -1,20 +1,4 @@
-/**
- * NativeScript Polyfills
- */
-
-// Install @nativescript/core polyfills (XHR, setTimeout, requestAnimationFrame)
-import '@nativescript/core/globals';
-// Install @nativescript/angular specific polyfills
-import '@nativescript/angular/polyfills';
-
-/**
- * Zone.js and patches
- */
-// Add pre-zone.js patches needed for the NativeScript platform
-import '@nativescript/zone-js/dist/pre-zone-polyfills';
-
-// Zone JS is required by default for Angular itself
-import 'zone.js';
-
-// Add NativeScript specific Zone JS patches
-import '@nativescript/zone-js';
+(window as any).global = window;
+(window as any).process = {
+  env: { DEBUG: undefined },
+};
